Hooked up new modal
