+++ src/shared/ModalContent.scss
@@ -23,25 +23,6 @@
   padding: 20px;
 }
 
-.data {
-  height: 40%;
-  width: 100%;
-  color: white;
-}
-
-.divider {
-  width: 95%;
-  margin: auto;
-  border-top: 1px solid rgba(255, 255, 255, .3);
-}
-
-.listsContainer {
-  height: 60%;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-}
-
 .imageContainer {
   display: flex;
   justify-content: center;
@@ -49,26 +30,4 @@
   width: 33%;
   height: 100%;
   padding: 30px;
-}
-
-.title {
-  font-size: 60px;
-  padding-bottom: 10px;
-}
-
-.bullet {
-  color: white;
-}
-
-.details {
-  display: flex;
-  flex-direction: row;
-  width: 260px;
-  justify-content: space-between;
-  padding-bottom: 30px;
-}
-
-.synopsis {
-  font-size: 15px;
-  line-height: 21px;
 }
