+++ src/shared/ModalDetails.scss
@@ -1,0 +1,46 @@
+
+.title {
+  font-size: 60px;
+  padding-bottom: 10px;
+}
+
+.bullet {
+  color: white;
+}
+
+.details {
+  display: flex;
+  flex-direction: row;
+  width: 260px;
+  justify-content: space-between;
+  padding-bottom: 30px;
+}
+
+.synopsis {
+  font-size: 15px;
+  line-height: 21px;
+}
+
+.data {
+  height: 40%;
+  width: 100%;
+  color: white;
+}
+
+.divider {
+  width: 95%;
+  margin: auto;
+  border-top: 1px solid rgba(255, 255, 255, .3);
+}
+
+.listsContainer {
+  height: 60%;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+}
+
+.content {
+  width: 100%;
+  height: 100%;
+}
