+++ src/shows/ShowModalContent.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+// import './ShowDetails.css';
+import PropTypes from 'prop-types';
+import ModalDetails from '../shared/ModalDetails';
+
+class ShowModalContent extends Component {
+
+  render() {
+    return <ModalDetails />
+  }
+}
+
+ShowModalContent.propTypes = {}
+
+export default ShowModalContent;
